requirements.txt
#Backend
fastapi
uvicorn
python-dotenv

#Database
sqlalchemy
psycopg2-binary
alembic

#Authentication
passlib[bcrypt]
python-jose[cryptography]
python-multipart
